Refactor remoteItemUpdated
